+++ client/lib/typings/abeamer.d.ts
@@ -22,6 +22,22 @@ declare namespace ABeamer {
   // ------------------------------------------------------------------------
 
   /**
+   * It simplifies the usage of [](VirtualAnimator) by plugins.
+   * In many cases plugins just need to receive the changing property,
+   * in order to modify its state.
+   * Override `animateProp` to receive the changing property.
+   * animateProp isn't called if the property is `uid`.
+   */
+  export interface SimpleVirtualAnimator {
+
+    props: AnyParams;
+    selector: string;
+    onAnimateProp: (name: PropName, value: PropValue) => void;
+    getProp(name: PropName): PropValue;
+    setProp(name: PropName, value: PropValue): void;
+  }
+
+  /**
    * Base class for all adapters: Element, Scene, Story,
    * and both DOM and virtual.
    */